BUVN-1.1/requirements.txt
torch>=2.0.0
transformers
sentencepiece
datasets
numpy
fastapi
uvicorn
pydantic
pyyaml
